Workaround for multiple header file paths
When defining a prebuilt library, the header file dirs seem to be
shared across all variations.

This example won't link, but if you run the compile tasks for
plat_32 and plat_64, you can see that all of the headers are
included in the compiler options in options.txt.

If you run the link tasks, you can see that the link options only
include the one library for the given platform.

I think this happens because all Prebuilt libraries share the
same underlying DefaultPrebuiltLibrary (which contains the
headers source set).

To workaround this, we hook into the AbstractNativeCompileTasks
and change the include paths before compiling.  We only cared
about C/C++, so I haven't checked if this would also work for
other languages (Obj-C/C++).